Add support for thing model function block features, including thing model APIs like CreateThingModel,UpdateThingModel,SetDeviceProperty,InvokeThingService etc.

+++ iot/include/alibabacloud/iot/model/RRpcResult.h
@@ -52,7 +52,7 @@ namespace AlibabaCloud
 				std::string errorMessage_;
 				std::string code_;
 				bool success_;
-				unsigned long long messageId_;
+				long messageId_;
 
 			};
 		}
